Initial commit
This is an advanced backbone.js tutorial application that uses HTML5,
jQuery and backbone.js. This is not a typical "Hello world" application.
You will learn the backbone.js in lightening quick speed with this
tutorial. Inside this tutorial, we are building a real time exam
question and answer style application. After you finish this tutorial,
you will get the black belt status in backbone.js. Download now to see
how you can master this latest technology in days not in weeks or
months. Please visit my blog @ www.backboneblackbelt.com that is
associated with this tutorial for more detailed information and in depth
explanation.